LPS-121804 Movee logic, no logic changes
